+++ consensus/hotstuff/vote_aggregator.go
@@ -2,28 +2,53 @@ package hotstuff
 
 import (
 	"github.com/onflow/flow-go/consensus/hotstuff/model"
-	"github.com/onflow/flow-go/model/flow"
 )
 
-// VoteAggregator aggregates votes and produces quorum certificates.
+// BlockSigner abstracts the implementation of how a signature of a block or a vote is produced
+// and stored in a stateful crypto object for aggregation.
+// The VoteAggregator implements both the VoteAggregator interface and the BlockSigner interface so that
+// the EventHandler could use the VoteAggregator interface to sign a Block, and Voter/BlockProducer can use
+// the BlockSigner interface to create vote.
+// When `CreateVote` is called, it internally creates stateful VoteCollector object, which also has the ability
+// to verify the block and generate the vote signature.
+// The created vote collector will be added to the vote collectors map. These
+// implementation details are abstracted to Voter/BlockProducer.
+type BlockSigner interface {
+	// CreateVote returns a vote for the given block.
+	// It returns:
+	//  - (vote, nil) if vote is created
+	//  - (nil , module.InvalidBlockError) if the block is invalid.
+	CreateVote(*model.Block) (*model.Vote, error)
+}
+
+// VoteAggregator verifies and aggregates votes to build QC.
+// When enough votes have been collected, it builds a QC and send it to the EventLoop
+// VoteAggregator also detects protocol violation, including invalid votes, double voting etc, and
+// notifies a HotStuff consumer for slashing.
 type VoteAggregator interface {
+	// AddVote verifies and aggregates a vote.
+	// The voting block could either be known or unknown.
+	// If the voting block is unknown, the vote won't be processed until AddBlock is called with the block.
+	// This method can be called concurrently, votes will be queued and processed asynchronously.
+	AddVote(vote *model.Vote)
 
-	// StorePendingVote is used to store a vote for a block for which we don't
-	// have a block yet.
-	StorePendingVote(vote *model.Vote) (bool, error)
+	// AddBlock notifies the VoteAggregator that it should start processing votes for the given block.
+	// AddBlock is a _synchronous_ call (logic is executed by the calling go routine). It also verifies
+	// validity of the proposer's vote for its own block.
+	// Expected error returns during normal operations:
+	// * model.InvalidBlockError if the proposer's vote for its own block is invalid
+	// * mempool.DecreasingPruningHeightError if the block's view has already been pruned
+	AddBlock(block *model.Proposal) error
 
-	// StoreVoteAndBuildQC will store a vote and build the QC related to the
-	// voted upon block if enough votes can be accumulated.
-	StoreVoteAndBuildQC(vote *model.Vote, block *model.Block) (*flow.QuorumCertificate, bool, error)
+	// InvalidBlock notifies the VoteAggregator about an invalid proposal, so that it
+	// can process votes for the invalid block and slash the voters. Expected error
+	// returns during normal operations:
+	// * mempool.DecreasingPruningHeightError if proposal's view has already been pruned
+	InvalidBlock(block *model.Proposal) error
 
-	// StoreProposerVote stores the vote of the proposer of the block and is
-	// used to separate vote and block handling.
-	StoreProposerVote(vote *model.Vote) bool
-
-	// BuildQCOnReceivedBlock will try to build a QC for the received block in
-	// case enough votes can be accumulated for it.
-	BuildQCOnReceivedBlock(block *model.Block) (*flow.QuorumCertificate, bool, error)
-
-	// PruneByView will remove any data held for the provided view.
-	PruneByView(view uint64)
+	// PruneUpToView prunes the vote collectors whose view is below the given view.
+	// If `view` is smaller than the previous value, the previous value is kept
+	// and no operations will be made. Sets highest pruned view which will be used
+	// to drop stale proposals and votes.
+	PruneUpToView(view uint64)
 }
